Remove deprecation warning during `npm install`
I realize this shouldn't be strictly necessary, because the loose semver range should pick up this version bump automatically.  However, this commit pins the coffee-script dep at 1.12.6 to make sure (especially with some of the NPM 5 edge case issues folks have been experiencing as of late).  It also locks to an exact dependency, for the same reasons as laid out here in this PR to the main `grunt` package: https://github.com/gruntjs/grunt/issues/1556#issuecomment-315189532